Refactor WeylExtraction class
This commit separates the WeylExtraction class into a family of
inherited classes as follows:

SurfaceExtraction <- SphericalExtraction <- WeylExtraction

The SurfaceExtraction class includes a member pointer to a
SurfaceGeometry abstract base class whose childs should be defined
appropriately for the relevant surface geometry. An example in the form
of SphericalGeometry is provided for the SphericalExtraction class.

There are also some other minor changes:
- The extraction_params_t struct in SimulationParametersBase.hpp is now
SphericalExtraction::params_t and there is some minor member variable
renaming
- The OpenMP parallelisation in `WeylExtraction::integrate_surface` has
been removed in `SurfaceExtraction::integrate`. If I can be persuaded
that there is a meaningful speed increase, then I will consider
reinstating it.
- Since you have the option of extracting derivatives of variables on
the surfaces, I have added a `Derivative::name` function to provide
a string for writing out depending on what type of derivative it is
- Probably some things I've forgotten about...

Note that only one small test has been done so more is required.